feat(FIN-039): create UiSlider component with customizable props and styles
