+++ src/client/features/tracking-operation/tracking-operation-filters/tracking-operation-sum-filter.tsx
@@ -4,7 +4,7 @@ import { FinControlledInput } from '@frontend/components/controlled-fields/fin-c
 import { cn } from '@frontend/shared/utils/cn.util';
 import type { FiltersDefaultProps } from '@frontend/features/tracking-operation/tracking-operation-filters/filters-default.props';
 import { NumberOnlyPattern } from '@common/constants/number-only-pattern.constant';
-import { Slider } from '@frontend/ui/ui-slider/slider';
+import { UiSlider } from '@frontend/ui/ui-slider/ui-slider';
 import { useFormContext, useWatch } from 'react-hook-form';
 
 //TODO ADD SLIDER BELOW INPUTS AND BIND IT'S VALUE TO INPUTS
@@ -43,7 +43,7 @@ export function SumFilter({ className }: FiltersDefaultProps) {
           pattern={NumberOnlyPattern}
         />
       </div>
-      <Slider
+      <UiSlider
         defaultValue={[startValue, endValue]}
         min={0}
         max={50000}
