+++ src/client/shared/ui/ui-slider/styles/slider-styles.scss
@@ -1,0 +1,114 @@
+@use 'sass:list';
+@import '../../../styles/variants';
+@import '../../../styles/functions';
+
+.slider {
+  position: relative;
+  display: flex;
+  width: 100%;
+  touch-action: none;
+  align-items: center;
+  user-select: none;
+
+  &[data-disabled] {
+    opacity: 0.5;
+  }
+
+  &[data-orientation='vertical'] {
+    height: 100%;
+    min-height: 11rem;
+    width: auto;
+    flex-direction: column;
+  }
+
+  @each $name, $colors in $variants {
+    &[data-variant='#{$name}'] {
+      --slider-color: #{list.nth($colors, 1)};
+    }
+  }
+
+  &[data-size='xs'] {
+    --slider-thumb-size: 0.625rem;
+    --slider-track-size: 0.125rem;
+  }
+
+  &[data-size='sm'] {
+    --slider-thumb-size: 0.75rem;
+    --slider-track-size: 0.1875rem;
+  }
+
+  &[data-size='default'] {
+    --slider-thumb-size: 1rem;
+    --slider-track-size: 0.375rem;
+  }
+
+  &[data-size='lg'] {
+    --slider-thumb-size: 1.25rem;
+    --slider-track-size: 0.5rem;
+  }
+
+  &[data-size='xl'] {
+    --slider-thumb-size: 1.5rem;
+    --slider-track-size: 0.625rem;
+  }
+}
+
+.slider-track {
+  position: relative;
+  flex-grow: 1;
+  overflow: hidden;
+  border-radius: 9999px;
+  background-color: var(--muted);
+
+  &[data-orientation='horizontal'] {
+    height: var(--slider-track-size, 0.375rem);
+    width: 100%;
+  }
+
+  &[data-orientation='vertical'] {
+    height: 100%;
+    width: var(--slider-track-size, 0.375rem);
+  }
+}
+
+.slider-range {
+  position: absolute;
+  background-color: var(--slider-color, var(--primary));
+
+  &[data-orientation='horizontal'] {
+    height: 100%;
+  }
+
+  &[data-orientation='vertical'] {
+    width: 100%;
+  }
+}
+
+.slider-thumb {
+  display: block;
+  flex-shrink: 0;
+  width: var(--slider-thumb-size, 1rem);
+  height: var(--slider-thumb-size, 1rem);
+  border-radius: 9999px;
+  border: 1px solid var(--slider-color, var(--primary));
+  background-color: white;
+  box-shadow: 0 1px 2px 0 rgb(0 0 0 / 5%);
+  outline: none;
+  transition:
+    color 0.15s,
+    box-shadow 0.15s;
+
+  &:hover {
+    box-shadow: 0 0 0 4px color-mix(in srgb, var(--slider-color, var(--primary)) 20%, transparent);
+  }
+
+  &:focus-visible {
+    box-shadow: 0 0 0 4px color-mix(in srgb, var(--slider-color, var(--primary)) 20%, transparent);
+  }
+
+  &:disabled {
+    pointer-events: none;
+    opacity: 0.5;
+  }
+}
+
